+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression/Json.hs
@@ -2,6 +2,8 @@
     DataKinds
   , FlexibleContexts
   , FlexibleInstances
+  , GADTs
+  , OverloadedLabels
   , OverloadedStrings
   , PolyKinds
   , TypeOperators
@@ -47,13 +49,39 @@ module Squeal.PostgreSQL.Expression.Json
   , jsonbSet
   , jsonbInsert
   , jsonbPretty
+    -- * Json and Jsonb set returning functions
+  , jsonEach
+  , jsonbEach
+  , jsonEachText
+  , jsonbEachText
+  , jsonObjectKeys
+  , jsonbObjectKeys
+  , jsonPopulateRecord
+  , jsonbPopulateRecord
+  , jsonPopulateRecordSet
+  , jsonbPopulateRecordSet
+  , jsonToRecord
+  , jsonbToRecord
+  , jsonToRecordSet
+  , jsonbToRecordSet
   ) where
 
+import Data.ByteString (ByteString)
+import GHC.TypeLits
+
+import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
+import Squeal.PostgreSQL.Expression.SetOf
+import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.List
+import Squeal.PostgreSQL.Query
+import Squeal.PostgreSQL.Render
 import Squeal.PostgreSQL.Schema
 
 import qualified Generics.SOP as SOP
+
+-- $setup
+-- >>> import Squeal.PostgreSQL
 
 {-----------------------------------------
  -- json and jsonb support
@@ -346,3 +374,151 @@ jsonbInsert tgt path val insertAfter = case insertAfter of
 -- | Returns its argument as indented JSON text.
 jsonbPretty :: null 'PGjsonb :--> null 'PGtext
 jsonbPretty = unsafeFunction "jsonb_pretty"
+
+
+-- | Expands the outermost JSON object into a set of key/value pairs.
+-- >>> import Data.Aeson
+-- >>> printSQL (select Star (from (jsonEach (literal (Json (object ["a" .= "foo", "b" .= "bar"]))))))
+-- SELECT * FROM json_each(('{"a":"foo","b":"bar"}' :: json))
+jsonEach
+  :: Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjson) -- ^ json object
+  -> FromClause outer commons schemas params
+      '["json_each" ::: '["key" ::: 'NotNull 'PGtext, "value" ::: 'NotNull 'PGjson]]
+jsonEach = unsafeSetOfFunction "json_each"
+
+-- | Expands the outermost binary JSON object into a set of key/value pairs.
+-- >>> import Data.Aeson
+-- >>> printSQL (select Star (from (jsonbEach (literal (Jsonb (object ["a" .= "foo", "b" .= "bar"]))))))
+-- SELECT * FROM jsonb_each(('{"a":"foo","b":"bar"}' :: jsonb))
+jsonbEach
+  :: Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjsonb) -- ^ jsonb object
+  -> FromClause outer commons schemas params
+      '["jsonb_each" ::: '["key" ::: 'NotNull 'PGtext, "value" ::: 'NotNull 'PGjson]]
+jsonbEach = unsafeSetOfFunction "jsonb_each"
+
+-- | Expands the outermost JSON object into a set of key/value pairs.-- | Expands the outermost binary JSON object into a set of key/value pairs.
+-- >>> import Data.Aeson
+-- >>> printSQL (select Star (from (jsonEachText (literal (Json (object ["a" .= "foo", "b" .= "bar"]))))))
+-- SELECT * FROM json_each_text(('{"a":"foo","b":"bar"}' :: json))
+jsonEachText
+  :: Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjson) -- ^ json object
+  -> FromClause outer commons schemas params
+      '["json_each" ::: '["key" ::: 'NotNull 'PGtext, "value" ::: 'NotNull 'PGtext]]
+jsonEachText = unsafeSetOfFunction "json_each_text"
+
+-- | Expands the outermost binary JSON object into a set of key/value pairs.
+-- >>> import Data.Aeson
+-- >>> printSQL (select Star (from (jsonbEachText (literal (Jsonb (object ["a" .= "foo", "b" .= "bar"]))))))
+-- SELECT * FROM jsonb_each_text(('{"a":"foo","b":"bar"}' :: jsonb))
+jsonbEachText
+  :: Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjsonb) -- ^ jsonb object
+  -> FromClause outer commons schemas params
+      '["jsonb_each" ::: '["key" ::: 'NotNull 'PGtext, "value" ::: 'NotNull 'PGtext]]
+jsonbEachText = unsafeSetOfFunction "jsonb_each_text"
+
+-- | Returns set of keys in the outermost JSON object.
+-- >>> import Data.Aeson
+-- >>> printSQL (jsonObjectKeys (literal (Json (object ["a" .= "foo", "b" .= "bar"]))))
+-- json_object_keys(('{"a":"foo","b":"bar"}' :: json))
+jsonObjectKeys
+  :: Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjson) -- ^ json object
+  -> FromClause outer commons schemas params
+      '["json_object_keys" ::: '["json_object_keys" ::: 'NotNull 'PGtext]]
+jsonObjectKeys = unsafeSetOfFunction "json_object_keys"
+
+-- | Returns set of keys in the outermost JSON object.
+-- >>> import Data.Aeson
+-- >>> printSQL (jsonbObjectKeys (literal (Jsonb (object ["a" .= "foo", "b" .= "bar"]))))
+-- jsonb_object_keys(('{"a":"foo","b":"bar"}' :: jsonb))
+jsonbObjectKeys
+  :: Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjsonb) -- ^ jsonb object
+  -> FromClause outer commons schemas params
+      '["jsonb_object_keys" ::: '["jsonb_object_keys" ::: 'NotNull 'PGtext]]
+jsonbObjectKeys = unsafeSetOfFunction "jsonb_object_keys"
+
+unsafePopulateFunction
+  :: KnownSymbol fun
+  => Alias fun
+  -> TypeExpression schemas (nullity ('PGcomposite row))
+  -> Expression outer commons 'Ungrouped schemas params '[]  ty
+  -> FromClause outer commons schemas params '[fun ::: row]
+unsafePopulateFunction fun ty expr = UnsafeFromClause $ renderSQL fun
+  <> parenthesized ("null::" <> renderSQL ty <> ", " <> renderSQL expr)
+
+-- | Expands the JSON expression to a row whose columns match the record
+-- type defined by the given table.
+jsonPopulateRecord
+  :: TypeExpression schemas (nullity ('PGcomposite row)) -- ^ row type
+  -> Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjson) -- ^ json object
+  -> FromClause outer commons schemas params '["json_populate_record" ::: row]
+jsonPopulateRecord = unsafePopulateFunction #json_populate_record
+
+-- | Expands the binary JSON expression to a row whose columns match the record
+-- type defined by the given table.
+jsonbPopulateRecord
+  :: TypeExpression schemas (nullity ('PGcomposite row)) -- ^ row type
+  -> Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjsonb) -- ^ jsonb object
+  -> FromClause outer commons schemas params '["jsonb_populate_record" ::: row]
+jsonbPopulateRecord = unsafePopulateFunction #jsonb_populate_record
+
+-- | Expands the outermost array of objects in the given JSON expression to a
+-- set of rows whose columns match the record type defined by the given table.
+jsonPopulateRecordSet
+  :: TypeExpression schemas (nullity ('PGcomposite row)) -- ^ row type
+  -> Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjson) -- ^ json array
+  -> FromClause outer commons schemas params '["json_populate_record_set" ::: row]
+jsonPopulateRecordSet = unsafePopulateFunction #json_populate_record_set
+
+-- | Expands the outermost array of objects in the given binary JSON expression
+-- to a set of rows whose columns match the record type defined by the given
+-- table.
+jsonbPopulateRecordSet
+  :: TypeExpression schemas (nullity ('PGcomposite row)) -- ^ row type
+  -> Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjsonb) -- ^ jsonb array
+  -> FromClause outer commons schemas params '["jsonb_populate_record_set" ::: row]
+jsonbPopulateRecordSet = unsafePopulateFunction #jsonb_populate_record_set
+
+unsafeRecordFunction
+  :: (SOP.SListI record, json `In` PGJsonType)
+  => ByteString
+  -> Expression outer commons 'Ungrouped schemas params '[] (nullity json)
+  -> Aliased (NP (Aliased (TypeExpression schemas))) (tab ::: record)
+  -> FromClause outer commons schemas params '[tab ::: record]
+unsafeRecordFunction fun expr (types `As` tab) = UnsafeFromClause $
+  fun <> parenthesized (renderSQL expr) <+> "AS" <+> renderSQL tab
+    <> parenthesized (renderCommaSeparated renderTy types)
+    where
+      renderTy :: Aliased (TypeExpression schemas) ty -> ByteString
+      renderTy (ty `As` alias) = renderSQL alias <+> renderSQL ty
+
+-- | Builds an arbitrary record from a JSON object.
+jsonToRecord
+  :: SOP.SListI record
+  => Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjson) -- ^ json object
+  -> Aliased (NP (Aliased (TypeExpression schemas))) (tab ::: record)
+  -> FromClause outer commons schemas params '[tab ::: record]
+jsonToRecord = unsafeRecordFunction "json_to_record"
+
+-- | Builds an arbitrary record from a binary JSON object.
+jsonbToRecord
+  :: SOP.SListI record
+  => Expression outer commons 'Ungrouped schemas params '[] (nullity 'PGjsonb) -- ^ jsonb object
+  -> Aliased (NP (Aliased (TypeExpression schemas))) (tab ::: record)
+  -> FromClause outer commons schemas params '[tab ::: record]
+jsonbToRecord = unsafeRecordFunction "jsonb_to_record"
+
+-- | Builds an arbitrary set of records from a JSON array of objects.
+jsonToRecordSet
+  :: SOP.SListI record
+  => Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjson) -- ^ json array
+  -> Aliased (NP (Aliased (TypeExpression schemas))) (tab ::: record)
+  -> FromClause outer commons schemas params '[tab ::: record]
+jsonToRecordSet = unsafeRecordFunction "json_to_record_set"
+
+-- | Builds an arbitrary set of records from a binary JSON array of objects.
+jsonbToRecordSet
+  :: SOP.SListI record
+  => Expression outer commons 'Ungrouped schemas params '[]  (nullity 'PGjsonb) -- ^ jsonb array
+  -> Aliased (NP (Aliased (TypeExpression schemas))) (tab ::: record)
+  -> FromClause outer commons schemas params '[tab ::: record]
+jsonbToRecordSet = unsafeRecordFunction "jsonb_to_record_set"
